requirements.txt
django==4.1
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
